enmap: fix area and fullsky_geometry functions.
The area() function now computes area precisely and quickly in a broad
class of geometries.  The fullsky_geometry function is more careful
with the RA axis native coordinate system branch cut; now all pixels
in the grid (that are not at the poles) have valid coordinates (across
their entire areas).