Handle pixel-alignment consistently when building cyl and zen WCS
Support the creation of zenithal WCS with a specified reference point,
but without pixel realignment.  This is achieved by adding argument
"align" to the wcsutils.build function.  The behavior of build and
enmap.geometry are unchanged for cylindrical projections*.  However,
we should consider deprecating build(..., ref='standard') in favor of
build(..., ref=None, align=True).

Footnote: geometry(..., ref='standard', force=True) produces a
different result than previously, even on cylindrical projections.
It's not clear if ref='standard' is valid input, however, and even if
it is, adding force=True to that should probably be invalid.