enmap: enable storing multiple maps in a single HDF5 file
The "address" argument to read_map, write_map, and any of the special
hdf functions will cause that address to be used as the group name in
the HDF5 file.

If you pass in an h5py Group object, instead of a filename, it will
read from or write to that handle (without trying to open a file).

When writing, with address=None, the output file is created or
overwritten if it exists.  When address is passed in, and that group
name already exists, it will be silently deleted prior to writing the
new data at that address.