Set default value of OMP_NUM_THREADS to zero (and not None)
The `DUCC0_NUM_THREADS` env. variable is converted to `long int` within
`ducc` (https://gitlab.mpcdf.mpg.de/mtr/ducc/-/blob/ducc0/src/ducc0/infra/threading.cc#L135) and
passing the string "None" is actually "well" converted to zero (so to leave `ducc` finds and
optimizes the number of threads) but it raises a kind of logic error (`Invalid argument`) since
"None" is not an integer. On laptop where `OMP_NUM_THREADS` is not necessary set, all goes well with
GNU C compiler (so no error) but the `clang` compiler (on mac for instance) is stricter and raises
an error.